refactor: shuttle route visualizer 컴포넌트 구조개선
